awsの呼び出しを別定義に修正
